pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-wiki"
version = "0.1.0"
description = "A Python-based, tool-agnostic Local LLM Knowledge Base and Memory Compiler."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "pyyaml",
    "pydantic",
    "python-dotenv",
    "markdown",
    "openai",
    "anthropic",
    "instructor",
    "groq",
    "python-frontmatter",
    "cffi",
    "cryptography",
    "markitdown[all]",
    "tenacity"
]

[project.scripts]
llm-wiki = "llm_wiki.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
exclude = ["venv", "venv.*"]
